Fixed the test suites
