+++ src/EditorFeatures/TestUtilities/ServicesTestUtilities.csproj
@@ -199,6 +199,7 @@
     <Compile Include="NavigateTo\NavigateToTestAggregator.Callback.cs" />
     <Compile Include="NavigateTo\NavigateToTestAggregator.cs" />
     <Compile Include="OptionExtensions.cs" />
+    <Compile Include="Preview\MockPreviewPaneService.cs" />
     <Compile Include="QuickInfo\AbstractQuickInfoSourceTests.cs" />
     <Compile Include="QuickInfo\AbstractSemanticQuickInfoSourceTests.cs" />
     <Compile Include="RenameTracking\MockPreviewDialogService.cs" />
